chore: UpdateApp component loading location
